requirements.txt
debugpy
opencv-python-headless
numpy
typer[all]
pillow